refactor tests part2
