+++ .env
@@ -1,17 +1,34 @@
-MONGODB_NAME=operandi_db
-MONGODB_USER=db_operandi
-MONGODB_PASS=db_operandi
-MONGODB_URL=mongodb://db_operandi:db_operandi@localhost:27017
-OPERANDI_HARVESTER_DEFAULT_USERNAME=harvester_operandi
-OPERANDI_HARVESTER_DEFAULT_PASSWORD=harvester_operandi
-OPERANDI_HPC_USERNAME=mmustaf
-OPERANDI_HPC_SSH_KEYPATH=/home/mm/.ssh/gwdg-cluster
-OPERANDI_HPC_PROJECT_NAME=operandi
+DOCKER_OPERANDI_NETWORK_NAME=operandi_network
+DOCKER_OPERANDI_NETWORK_MTU=1450
+
 OPERANDI_LOGS_DIR=/tmp/operandi_logs
-OPERANDI_RABBITMQ_CONFIG_JSON=./src/rabbitmq_definitions.json
-RABBITMQ_URL=amqp://operandi_user:operandi_password@localhost:5672/
 OPERANDI_SERVER_BASE_DIR=/tmp/operandi_data
-OPERANDI_SERVER_DEFAULT_USERNAME=server_operandi
-OPERANDI_SERVER_DEFAULT_PASSWORD=server_operandi
-OPERANDI_SERVER_URL_LIVE=http://localhost:8000
-OPERANDI_SERVER_URL_LOCAL=http://0.0.0.0:8000
+
+OPERANDI_HARVESTER_DEFAULT_USERNAME=operandi
+OPERANDI_HARVESTER_DEFAULT_PASSWORD=operandi
+
+OPERANDI_HPC_USERNAME=mmustaf
+OPERANDI_HPC_PROJECT_NAME=operandi
+OPERANDI_HPC_SSH_KEYPATH=/home/root/.ssh/gwdg_hpc_key
+
+OPERANDI_SERVER_DEFAULT_USERNAME=operandi
+OPERANDI_SERVER_DEFAULT_PASSWORD=operandi
+OPERANDI_SERVER_HOST=operandi-server-docker-host
+OPERANDI_SERVER_PORT=8000
+OPERANDI_SERVER_URL=http://${OPERANDI_SERVER_HOST}.${DOCKER_OPERANDI_NETWORK_NAME}:${OPERANDI_SERVER_PORT}
+
+MONGODB_NAME=operandi_db
+MONGODB_USER=operandi
+MONGODB_PASS=operandi
+MONGODB_HOST=operandi-mongodb-docker-host
+MONGODB_PORT=27017
+MONGODB_URL=mongodb://${MONGODB_USER}:${MONGODB_PASS}@${MONGODB_HOST}.${DOCKER_OPERANDI_NETWORK_NAME}:${MONGODB_PORT}
+
+RABBITMQ_FEATURE_FLAGS=quorum_queue,implicit_default_bindings,classic_mirrored_queue_version
+RABBITMQ_USER=operandi
+RABBITMQ_PASS=operandi
+RABBITMQ_HOST=operandi-rabbitmq-docker-host
+RABBITMQ_PORT=5672
+RABBITMQ_PORT_UI=15672
+RABBITMQ_PORT_CLI=25672
+RABBITMQ_URL=amqp://${RABBITMQ_USER}:${RABBITMQ_PASS}@${RABBITMQ_HOST}.${DOCKER_OPERANDI_NETWORK_NAME}:${RABBITMQ_PORT}
