--- docker.env
@@ -1,17 +1,0 @@
-OPERANDI_DB_NAME=operandi_db
-OPERANDI_DB_ROOT_USER=db_operandi
-OPERANDI_DB_ROOT_PASS=db_operandi
-OPERANDI_DB_URL=mongodb://db_operandi:db_operandi@mongo-db-host:27017
-OPERANDI_HARVESTER_DEFAULT_USERNAME=harvester_operandi
-OPERANDI_HARVESTER_DEFAULT_PASSWORD=harvester_operandi
-OPERANDI_HPC_USERNAME=mmustaf
-OPERANDI_HPC_SSH_KEYPATH=/home/mm/.ssh/gwdg-cluster
-OPERANDI_HPC_PROJECT_NAME=operandi
-OPERANDI_LOGS_DIR=/tmp/operandi_logs
-OPERANDI_RABBITMQ_CONFIG_JSON=./src/rabbitmq_definitions.json
-OPERANDI_RABBITMQ_URL=amqp://operandi_user:operandi_password@rabbit-mq-host:5672/
-OPERANDI_SERVER_BASE_DIR=/tmp/operandi_data
-OPERANDI_SERVER_DEFAULT_USERNAME=server_operandi
-OPERANDI_SERVER_DEFAULT_PASSWORD=server_operandi
-OPERANDI_SERVER_URL_LIVE=http://localhost:8000
-OPERANDI_SERVER_URL_LOCAL=http://0.0.0.0:8000
