+++ tests/.env
@@ -1,7 +1,8 @@
-DOCKER_OPERANDI_NETWORK_NAME=operandi_test_network
+DOCKER_OPERANDI_NETWORK_NAME=operandi_network
 DOCKER_OPERANDI_NETWORK_MTU=1450
 
-OPERANDI_LOGS_DIR=/tmp/operandi_tests_logs
+OPERANDI_LOGS_DIR=/tmp/operandi_logs_test
+OPERANDI_SERVER_BASE_DIR=/tmp/operandi_data_test
 
 OPERANDI_HARVESTER_DEFAULT_USERNAME=operandi_test
 OPERANDI_HARVESTER_DEFAULT_PASSWORD=operandi_test
@@ -10,24 +11,23 @@ OPERANDI_HPC_USERNAME=mmustaf
 OPERANDI_HPC_PROJECT_NAME=operandi_tests
 OPERANDI_HPC_SSH_KEYPATH=/home/root/.ssh/gwdg_hpc_key
 
-OPERANDI_SERVER_BASE_DIR=/tmp/ocrd_tests_data
 OPERANDI_SERVER_DEFAULT_USERNAME=operandi_test
 OPERANDI_SERVER_DEFAULT_PASSWORD=operandi_test
-OPERANDI_SERVER_HOST=operandi-server-test-docker-host
+OPERANDI_SERVER_HOST=operandi-server-docker-host-test
 OPERANDI_SERVER_PORT=8000
 OPERANDI_SERVER_URL=http://${OPERANDI_SERVER_HOST}.${DOCKER_OPERANDI_NETWORK_NAME}:${OPERANDI_SERVER_PORT}
 
-MONGODB_NAME=operandi_db
+MONGODB_NAME=operandi_db_test
 MONGODB_USER=operandi_test
 MONGODB_PASS=operandi_test
-MONGODB_HOST=mongodb-test-docker-host
+MONGODB_HOST=operandi-mongodb-docker-host-test
 MONGODB_PORT=27017
 MONGODB_URL=mongodb://${MONGODB_USER}:${MONGODB_PASS}@${MONGODB_HOST}.${DOCKER_OPERANDI_NETWORK_NAME}:${MONGODB_PORT}
 
 RABBITMQ_FEATURE_FLAGS=quorum_queue,implicit_default_bindings,classic_mirrored_queue_version
 RABBITMQ_USER=operandi_test
 RABBITMQ_PASS=operandi_test
-RABBITMQ_HOST=rabbitmq-test-docker-host
+RABBITMQ_HOST=operandi-rabbitmq-docker-host-test
 RABBITMQ_PORT=5672
 RABBITMQ_PORT_UI=15672
 RABBITMQ_PORT_CLI=25672
