--- tests/test_integration_5_harvester.py
@@ -1,2 +1,0 @@
-def test_dummy_harvester():
-    pass
