+++ tests/test_integration_5_server_3_workspace.py
@@ -1,19 +1,32 @@
 from os import environ
 from os.path import join
-from operandi_server.constants import SERVER_WORKSPACES_ROUTER
+from requests import (
+    delete as requests_delete,
+    get as requests_get,
+    post as requests_post,
+    put as requests_put
+)
+from operandi_utils.constants import StateWorkspace
 from operandi_utils.database import sync_db_get_workspace, sync_db_update_workspace
-from operandi_utils import StateWorkspace
+from operandi_server.constants import SERVER_WORKSPACES_ROUTER
+from tests.helpers.asserts import (
+    assert_exists_ws_db_resource,
+    assert_exists_ws_db_resource_not,
+    assert_local_dir_workspace,
+    assert_local_dir_workspace_not
+)
 
-from tests.helpers_asserts import assert_exists_ws_db_resource, assert_exists_ws_db_resource_not
-from .helpers_asserts import assert_local_dir_workspace, assert_local_dir_workspace_not
+OPERANDI_SERVER_URL = environ.get("OPERANDI_SERVER_URL")
 
-"""
-def test_post_workspace_url(operandi, auth):
+
+def test_post_workspace_url(auth):
     mets_url = "https://content.staatsbibliothek-berlin.de/dc/PPN631277528.mets.xml"
     # Separate with `,` to add a second file group to be preserved, e.g., `DEFAULT,MAX`
     preserve_file_grps = "DEFAULT"
-    request_url = f"/import_external_workspace?mets_url={mets_url}&preserve_file_grps={preserve_file_grps}"
-    response = operandi.post(url=request_url, auth=auth)
+    request_url = f"{OPERANDI_SERVER_URL}/import_external_workspace"
+    request_url += f"?mets_url={mets_url}"
+    request_url += f"&preserve_file_grps={preserve_file_grps}"
+    response = requests_post(url=request_url, auth=auth)
     assert response.status_code == 201
     ws_id = response.json()["resource_id"]
     ws_db = sync_db_get_workspace(workspace_id=ws_id)
@@ -21,9 +34,9 @@ def test_post_workspace_url(operandi, auth):
     assert_local_dir_workspace(ws_id)
 
 
-def test_post_workspace_zip(operandi, auth, bytes_dummy_workspace):
+def test_post_workspace_zip(auth, bytes_dummy_workspace):
     ws_file = {"workspace": bytes_dummy_workspace}
-    response = operandi.post(url="/workspace", files=ws_file, auth=auth)
+    response = requests_post(url=f"{OPERANDI_SERVER_URL}/workspace", files=ws_file, auth=auth)
     assert response.status_code == 201
     ws_id = response.json()["resource_id"]
     ws_db = sync_db_get_workspace(workspace_id=ws_id)
@@ -31,9 +44,9 @@ def test_post_workspace_zip(operandi, auth, bytes_dummy_workspace):
     assert_local_dir_workspace(ws_id)
 
 
-def test_post_workspace_zip_different_mets(operandi, auth, bytes_ws_different_mets):
+def test_post_workspace_zip_different_mets(auth, bytes_ws_different_mets):
     ws_file = {"workspace": bytes_ws_different_mets}
-    response = operandi.post(url="/workspace", files=ws_file, auth=auth)
+    response = requests_post(url=f"{OPERANDI_SERVER_URL}/workspace", files=ws_file, auth=auth)
     assert response.status_code == 201
     ws_id = response.json()["resource_id"]
     ws_db = sync_db_get_workspace(workspace_id=ws_id)
@@ -41,11 +54,11 @@ def test_post_workspace_zip_different_mets(operandi, auth, bytes_ws_different_me
     assert_local_dir_workspace(ws_id)
 
 
-def test_put_workspace_zip(operandi, auth, bytes_dummy_workspace, bytes_ws_different_mets):
+def test_put_workspace_zip(auth, bytes_dummy_workspace, bytes_ws_different_mets):
     put_ws_id = "put_workspace_id"
     # The first put request creates a new workspace
     ws_file = {"workspace": bytes_dummy_workspace}
-    response = operandi.put(url=f"/workspace/{put_ws_id}", files=ws_file, auth=auth)
+    response = requests_put(url=f"{OPERANDI_SERVER_URL}/workspace/{put_ws_id}", files=ws_file, auth=auth)
     assert response.status_code == 201
     ws_id = response.json()["resource_id"]
     ws_db = sync_db_get_workspace(workspace_id=ws_id)
@@ -57,7 +70,7 @@ def test_put_workspace_zip(operandi, auth, bytes_dummy_workspace, bytes_ws_diffe
 
     # The second put request replaces the previously created workspace
     ws_file_diff = {"workspace": bytes_ws_different_mets}
-    response = operandi.put(f"/workspace/{put_ws_id}", files=ws_file_diff, auth=auth)
+    response = requests_put(url=f"{OPERANDI_SERVER_URL}/workspace/{put_ws_id}", files=ws_file_diff, auth=auth)
     assert response.status_code == 201
     ws_id = response.json()["resource_id"]
 
@@ -71,10 +84,10 @@ def test_put_workspace_zip(operandi, auth, bytes_dummy_workspace, bytes_ws_diffe
         f"Ocrd identifiers should not, but match: {ocrd_identifier1} == {ocrd_identifier2}"
 
 
-def test_delete_workspace(operandi, auth, bytes_ws_different_mets):
+def test_delete_workspace(auth, bytes_ws_different_mets):
     # Post a workspace
     ws_file = {"workspace": bytes_ws_different_mets}
-    response = operandi.post(url="/workspace", files=ws_file, auth=auth)
+    response = requests_post(url=f"{OPERANDI_SERVER_URL}/workspace", files=ws_file, auth=auth)
     assert response.status_code == 201
     posted_ws_id = response.json()["resource_id"]
 
@@ -84,29 +97,29 @@ def test_delete_workspace(operandi, auth, bytes_ws_different_mets):
 
     # Delete the previously posted workspace
     delete_ws_id = posted_ws_id
-    response = operandi.delete(f"/workspace/{delete_ws_id}", auth=auth)
+    response = requests_delete(url=f"{OPERANDI_SERVER_URL}/workspace/{delete_ws_id}", auth=auth)
     assert response.status_code == 200
     ws_db_deleted = sync_db_get_workspace(workspace_id=posted_ws_id)
     assert_exists_ws_db_resource_not(ws_db_deleted, workspace_id=delete_ws_id)
     assert_local_dir_workspace_not(delete_ws_id)
 
 
-def test_delete_workspace_non_existing(operandi, auth, bytes_ws_different_mets):
+def test_delete_workspace_non_existing(auth, bytes_ws_different_mets):
     ws_file = {"workspace": bytes_ws_different_mets}
-    response = operandi.post("/workspace", files=ws_file, auth=auth)
+    response = requests_post(url=f"{OPERANDI_SERVER_URL}/workspace", files=ws_file, auth=auth)
     posted_ws_id = response.json()["resource_id"]
     delete_ws_id = posted_ws_id
-    response = operandi.delete(f"/workspace/{delete_ws_id}", auth=auth)
+    response = requests_delete(url=f"{OPERANDI_SERVER_URL}/workspace/{delete_ws_id}", auth=auth)
     assert response.status_code == 200  # delete success
-    response = operandi.delete(f"/workspace/{delete_ws_id}", auth=auth)
+    response = requests_delete(url=f"{OPERANDI_SERVER_URL}/workspace/{delete_ws_id}", auth=auth)
     assert response.status_code == 410  # already deleted
 
 
-def test_get_workspace(operandi, auth, bytes_ws_different_mets):
+def test_get_workspace(auth, bytes_ws_different_mets):
     ws_file = {"workspace": bytes_ws_different_mets}
-    response = operandi.post("/workspace", files=ws_file, auth=auth)
+    response = requests_post(url=f"{OPERANDI_SERVER_URL}/workspace", files=ws_file, auth=auth)
     ws_id = response.json()["resource_id"]
-    response = operandi.get(f"/workspace/{ws_id}", auth=auth)
+    response = requests_get(url=f"{OPERANDI_SERVER_URL}/workspace/{ws_id}", auth=auth)
     assert response.status_code == 200
     print(response.headers)
     assert response.headers.get("content-type").find("zip") > -1, "content-type should be something with 'zip'"
@@ -123,15 +136,15 @@ def test_get_workspace(operandi, auth, bytes_ws_different_mets):
                 filePtr.write(chunk)
 
 
-def test_get_workspace_non_existing(operandi, auth):
+def test_get_workspace_non_existing(auth):
     non_ws_id = "non_existing_workspace_id"
-    response = operandi.get(f"/workspace/{non_ws_id}", auth=auth)
+    response = requests_get(url=f"{OPERANDI_SERVER_URL}/workspace/{non_ws_id}", auth=auth)
     assert response.status_code == 404
 
 
-def test_get_not_ready_workspace(operandi, auth, bytes_dummy_workspace):
+def test_get_not_ready_workspace(auth, bytes_dummy_workspace):
     ws_file = {"workspace": bytes_dummy_workspace}
-    response = operandi.post("/workspace", files=ws_file, auth=auth)
+    response = requests_post(url=f"{OPERANDI_SERVER_URL}/workspace", files=ws_file, auth=auth)
     assert response.status_code == 201
     ws_id = response.json()["resource_id"]
 
@@ -140,6 +153,5 @@ def test_get_not_ready_workspace(operandi, auth, bytes_dummy_workspace):
     assert_local_dir_workspace(ws_id)
 
     sync_db_update_workspace(find_workspace_id=ws_id, state=StateWorkspace.TRANSFERRING_TO_HPC)
-    response = operandi.get(f"/workspace/{ws_id}", auth=auth)
+    response = requests_get(url=f"{OPERANDI_SERVER_URL}/workspace/{ws_id}", auth=auth)
     assert response.status_code == 403  # State not READY
-"""
