+++ tests/test_integration_6_harvester.py
@@ -1,30 +1,26 @@
 from os import environ
 from os.path import join
+from requests import get as requests_get, post as requests_post
 from time import sleep
 
 from operandi_utils.constants import StateJob
-from operandi_utils.rabbitmq import RABBITMQ_QUEUE_HARVESTER, RABBITMQ_QUEUE_JOB_STATUSES
 
-"""
+OPERANDI_SERVER_URL = environ.get("OPERANDI_SERVER_URL")
 
-def test_full_cycle(auth_harvester, service_broker, bytes_default_workflow, bytes_small_workspace):
-    response = operandi.get("/")
+
+def test_full_cycle(auth_harvester, bytes_default_workflow, bytes_small_workspace):
+    response = requests_get(f"{OPERANDI_SERVER_URL}/")
     assert response.json()["message"] == "The home page of the OPERANDI Server"
-
-    # Create a background worker for the harvester queue
-    service_broker.create_worker_process(queue_name=RABBITMQ_QUEUE_HARVESTER, status_checker=False)
-    # Create a background worker for the job statuses queue
-    service_broker.create_worker_process(queue_name=RABBITMQ_QUEUE_JOB_STATUSES, status_checker=True)
 
     # Post a workflow script
     wf_file = {"nextflow_script": bytes_default_workflow}
-    response = operandi.post(url="/workflow", files=wf_file, auth=auth_harvester)
+    response = requests_post(url=f"{OPERANDI_SERVER_URL}/workflow", files=wf_file, auth=auth_harvester)
     assert response.status_code == 201
     workflow_id = response.json()["resource_id"]
 
     # Post a workspace zip
     ws_file = {"workspace": bytes_small_workspace}
-    response = operandi.post(url="/workspace", files=ws_file, auth=auth_harvester)
+    response = requests_post(url=f"{OPERANDI_SERVER_URL}/workspace", files=ws_file, auth=auth_harvester)
     assert response.status_code == 201
     workspace_id = response.json()["resource_id"]
 
@@ -42,7 +38,7 @@ def test_full_cycle(auth_harvester, service_broker, bytes_default_workflow, byte
           "ram": 32
         }
     }
-    response = operandi.post(url=f"/workflow/{workflow_id}", json=req_data, auth=auth_harvester)
+    response = requests_post(url=f"{OPERANDI_SERVER_URL}/workflow/{workflow_id}", json=req_data, auth=auth_harvester)
     assert response.status_code == 201
     workflow_job_id = response.json()["resource_id"]
 
@@ -51,7 +47,9 @@ def test_full_cycle(auth_harvester, service_broker, bytes_default_workflow, byte
     while tries > 0:
         tries -= 1
         sleep(30)
-        response = operandi.get(url=f"/workflow/{workflow_id}/{workflow_job_id}", auth=auth_harvester)
+        response = requests_get(
+            url=f"{OPERANDI_SERVER_URL}/workflow/{workflow_id}/{workflow_job_id}", auth=auth_harvester
+        )
         assert response.status_code == 200
         job_status = response.json()["job_state"]
         if job_status == StateJob.SUCCESS:
@@ -68,10 +66,11 @@ def test_full_cycle(auth_harvester, service_broker, bytes_default_workflow, byte
     # TODO: Fix this, wait for 10 secs till
     #  the data is transferred from HPC to Operandi Server
     sleep(10)
-    response = operandi.get(url=f"/workflow/{workflow_id}/{workflow_job_id}/log", auth=auth_harvester)
+    response = requests_get(
+        url=f"{OPERANDI_SERVER_URL}/workflow/{workflow_id}/{workflow_job_id}/log", auth=auth_harvester
+    )
     zip_local_path = join(environ.get("OPERANDI_SERVER_BASE_DIR"), f"{workflow_job_id}.zip")
     with open(zip_local_path, "wb") as filePtr:
         for chunk in response.iter_bytes(chunk_size=1024):
             if chunk:
                 filePtr.write(chunk)
-"""
