--- tests/test_integration_6_server.py
@@ -1,2 +1,0 @@
-def test_dummy_server():
-    pass
