--- tests/tests_harvester/test_dummy.py
@@ -1,2 +1,0 @@
-def test_dummy():
-    pass
