--- tests/tests_server/test_endpoint_root.py
@@ -1,6 +1,0 @@
-"""
-def test_get_root_info(operandi):
-    response = operandi.get("/")
-    assert response.status_code == 200
-    assert response.json()["message"] == "The home page of the OPERANDI Server"
-"""
